Added to UI wall, Added bag, Added bin
Stage changes:

Made a new trashbin with an CSG nodes, added a built in bag to the scene.

Bag script changes:

Created a new script for the bag, it detects when the right trash has been thrown into it by the player.

stage script changes:

Changed the limit for each liquid waste and trash waste