Added UI script, edited bag script
Added UI script, inside it has signals connected to the bag, when the bag is filled with content then the bar is increased.

added a place holder signal for increasing the bag limit.

In the bag script the bag can only take in more trash if is not full, and then emits a signal every time it takes in a new value.
Made 2 signals for the UI