Edited bin script, added character script
Changed trashbins layer and mask to read in the bag and trash.

bin script now has 2 signals set up for adding points and emptying the bag, add points will be sent to the character script, which will then hold the number of points and this value will increase.

The signal empty bag will be sent to the bag to clear its content setting it to 0 and setting the progress bar back down to 0.